test(l1): move ethrex namespace tests to the integration test crate
