+++ crates/networking/rpc/ethrex.rs
@@ -30,9 +30,9 @@ use crate::{
 #[derive(Debug)]
 pub struct SimulateFrameTransactionRequest {
     /// Decoded type-`0x06` frame transaction (validated in `parse`).
-    transaction: Transaction,
+    pub transaction: Transaction,
     /// Block the simulation runs against. Defaults to `latest`.
-    block: Option<BlockIdentifierOrHash>,
+    pub block: Option<BlockIdentifierOrHash>,
 }
 
 /// Result of `ethrex_simulateFrameTransaction`.
@@ -328,75 +328,4 @@ fn to_hex_u256(value: U256) -> String {
 
 fn to_value(result: SimulateFrameTransactionResult) -> Result<Value, RpcErr> {
     serde_json::to_value(result).map_err(|error| RpcErr::Internal(error.to_string()))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ethrex_common::types::{EIP1559Transaction, FrameTransaction};
-    use serde_json::json;
-
-    /// Canonical (`type || payload`) hex, `0x`-prefixed, for a transaction.
-    fn raw_hex(tx: &Transaction) -> String {
-        let mut buf = Vec::new();
-        tx.encode_canonical(&mut buf);
-        format!("0x{}", hex::encode(buf))
-    }
-
-    #[test]
-    fn parse_accepts_frame_tx_without_block() {
-        let tx = Transaction::FrameTransaction(FrameTransaction::default());
-        let params = Some(vec![json!(raw_hex(&tx))]);
-        let parsed = SimulateFrameTransactionRequest::parse(&params).expect("frame tx accepted");
-        assert!(matches!(
-            parsed.transaction,
-            Transaction::FrameTransaction(_)
-        ));
-        assert!(parsed.block.is_none());
-    }
-
-    #[test]
-    fn parse_accepts_optional_block_tag() {
-        let tx = Transaction::FrameTransaction(FrameTransaction::default());
-        let params = Some(vec![json!(raw_hex(&tx)), json!("latest")]);
-        let parsed = SimulateFrameTransactionRequest::parse(&params).expect("frame tx accepted");
-        assert!(parsed.block.is_some());
-    }
-
-    #[test]
-    fn parse_rejects_non_frame_tx() {
-        let tx = Transaction::EIP1559Transaction(EIP1559Transaction::default());
-        let params = Some(vec![json!(raw_hex(&tx))]);
-        let err = SimulateFrameTransactionRequest::parse(&params).unwrap_err();
-        assert!(matches!(err, RpcErr::BadParams(msg) if msg.contains("frame")));
-    }
-
-    #[test]
-    fn parse_rejects_missing_0x_prefix() {
-        let params = Some(vec![json!("abcdef")]);
-        let err = SimulateFrameTransactionRequest::parse(&params).unwrap_err();
-        assert!(matches!(err, RpcErr::BadParams(_)));
-    }
-
-    #[test]
-    fn parse_rejects_empty_and_missing_params() {
-        assert!(matches!(
-            SimulateFrameTransactionRequest::parse(&Some(vec![])),
-            Err(RpcErr::BadParams(_))
-        ));
-        assert!(matches!(
-            SimulateFrameTransactionRequest::parse(&None),
-            Err(RpcErr::BadParams(_))
-        ));
-    }
-
-    #[test]
-    fn parse_rejects_too_many_params() {
-        let tx = Transaction::FrameTransaction(FrameTransaction::default());
-        let params = Some(vec![json!(raw_hex(&tx)), json!("latest"), json!("extra")]);
-        assert!(matches!(
-            SimulateFrameTransactionRequest::parse(&params),
-            Err(RpcErr::BadParams(_))
-        ));
-    }
 }
