+++ crates/networking/rpc/lib.rs
@@ -64,7 +64,7 @@ mod authentication;
 pub mod debug;
 pub mod engine;
 mod eth;
-mod ethrex;
+pub mod ethrex;
 mod mempool;
 mod net;
 pub mod rpc;
