+++ crates/networking/rpc/utils.rs
@@ -461,28 +461,3 @@ pub fn parse_json_hex(hex: &serde_json::Value) -> Result<u64, String> {
         Err(format!("Could not parse given hex {hex}"))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn from_prefix_recognizes_ethrex_namespace() {
-        assert_eq!(
-            RpcNamespace::from_prefix("ethrex"),
-            Some(RpcNamespace::Ethrex)
-        );
-    }
-
-    #[test]
-    fn ethrex_method_resolves_to_ethrex_namespace() {
-        let req = RpcRequest::new("ethrex_simulateFrameTransaction", None);
-        assert_eq!(req.namespace().unwrap(), RpcNamespace::Ethrex);
-    }
-
-    #[test]
-    fn unknown_namespace_is_method_not_found() {
-        let req = RpcRequest::new("bogus_method", None);
-        assert!(matches!(req.namespace(), Err(RpcErr::MethodNotFound(_))));
-    }
-}
