+++ test/Cargo.toml
@@ -48,6 +48,7 @@ lazy_static.workspace = true
 bytes.workspace = true
 serde_json.workspace = true
 ethrex.workspace = true
+clap.workspace = true
 ethrex-l2-common.workspace = true
 ethrex-sdk.workspace = true
 # Enable SQL for tests so we don't need `cargo test -p ethrex-test --features sql`.
