+++ test/tests/cmd/mod.rs
@@ -1,1 +1,2 @@
+mod cli_tests;
 mod decode_tests;
