+++ test/tests/rpc/mod.rs
@@ -3,5 +3,7 @@ mod block_access_list_tests;
 mod client_version_tests;
 mod fork_choice_tests;
 mod http_batch_tests;
+mod namespace_tests;
 mod send_raw_transaction_tests;
+mod simulate_frame_transaction_tests;
 mod subscription_manager_tests;
